chore: delete model classes that were moved to EDC
