+++ core/identity-hub-api/build.gradle.kts
@@ -22,11 +22,11 @@ dependencies {
     api(libs.edc.spi.jsonld)
     api(libs.edc.spi.core)
     api(project(":spi:identity-hub-spi"))
-    implementation(project(":core:identity-hub-transform"))
     implementation(libs.edc.spi.validator)
     implementation(libs.edc.spi.web)
     implementation(libs.edc.core.jerseyproviders)
     implementation(libs.edc.core.transform)
+    implementation(libs.edc.iatp.transform)
     implementation(libs.jakarta.rsApi)
     testImplementation(libs.edc.junit)
     testImplementation(libs.edc.ext.jsonld)
