+++ core/identity-hub-api/src/test/java/org/eclipse/edc/identityservice/api/validation/PresentationQueryValidatorTest.java
@@ -19,11 +19,11 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.json.JsonArray;
 import jakarta.json.JsonObject;
-import org.eclipse.edc.identityhub.spi.model.PresentationQuery;
-import org.eclipse.edc.identityhub.spi.model.presentationdefinition.Constraints;
-import org.eclipse.edc.identityhub.spi.model.presentationdefinition.Field;
-import org.eclipse.edc.identityhub.spi.model.presentationdefinition.InputDescriptor;
-import org.eclipse.edc.identityhub.spi.model.presentationdefinition.PresentationDefinition;
+import org.eclipse.edc.identitytrust.model.credentialservice.PresentationQuery;
+import org.eclipse.edc.identitytrust.model.presentationdefinition.Constraints;
+import org.eclipse.edc.identitytrust.model.presentationdefinition.Field;
+import org.eclipse.edc.identitytrust.model.presentationdefinition.InputDescriptor;
+import org.eclipse.edc.identitytrust.model.presentationdefinition.PresentationDefinition;
 import org.eclipse.edc.jsonld.TitaniumJsonLd;
 import org.eclipse.edc.jsonld.spi.JsonLd;
 import org.eclipse.edc.jsonld.spi.JsonLdKeywords;
