--- core/identity-hub-transform/build.gradle.kts
@@ -1,28 +1,0 @@
-/*
- *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-plugins {
-    `java-library`
-    `maven-publish`
-}
-
-dependencies {
-    api(project(":spi:identity-hub-spi"))
-    api(libs.edc.spi.jsonld)
-    api(libs.edc.spi.transform)
-
-    testImplementation(libs.edc.ext.jsonld)
-    testImplementation(libs.edc.core.transform)
-    testImplementation(libs.edc.junit)
-}
