--- core/identity-hub-transform/src/test/resources/presentation_ex.json
@@ -1,15 +1,0 @@
-{
-  "@context": {
-    "@version": 1.1,
-    "PresentationSubmission": {
-      "@id": "https://identity.foundation/presentation-exchange/#presentation-submission",
-      "@context": {
-        "@version": 1.1,
-        "presentation_submission": {
-          "@id": "https://identity.foundation/presentation-exchange/#presentation-submission",
-          "@type": "@json"
-        }
-      }
-    }
-  }
-}
