--- core/identity-hub-transform/src/test/resources/presentation_query.json
@@ -1,12 +1,0 @@
-{
-  "@context": {
-    "scope": {
-      "@id": "https://w3id.org/tractusx-trust/v0.8/scope",
-      "@container": "@set"
-    },
-    "presentationDefinition": {
-      "@id": "https://w3id.org/tractusx-trust/v0.8/presentationDefinition",
-      "@type": "@json"
-    }
-  }
-}
