+++ gradle/libs.versions.toml
@@ -49,6 +49,7 @@ edc-identity-did-core = { module = "org.eclipse.edc:identity-did-core", version.
 edc-identity-did-web = { module = "org.eclipse.edc:identity-did-web", version.ref = "edc" }
 edc-iatp-service = { module = "org.eclipse.edc:identity-trust-service", version.ref = "edc" }
 edc-iatp-core = { module = "org.eclipse.edc:identity-trust-core", version.ref = "edc" }
+edc-iatp-transform = { module = "org.eclipse.edc:identity-trust-transform", version.ref = "edc" }
 edc-vc-jwt = { module = "org.eclipse.edc:jwt-verifiable-credentials", version.ref = "edc" }
 edc-vc-ldp = { module = "org.eclipse.edc:ldp-verifiable-credentials", version.ref = "edc" }
 edc-ext-http = { module = "org.eclipse.edc:http", version.ref = "edc" }
