+++ spi/identity-hub-spi/src/main/java/org/eclipse/edc/identityhub/spi/model/IdentityHubConstants.java
@@ -14,15 +14,7 @@
 
 package org.eclipse.edc.identityhub.spi.model;
 
-public interface IdentityHubConstants {
-    String IATP_CONTEXT_URL = "https://w3id.org/tractusx-trust/v0.8";
-    String IATP_PREFIX = IATP_CONTEXT_URL + "/";
-    String PRESENTATION_EXCHANGE_URL = "https://identity.foundation/presentation-exchange/submission/v1";
-    String W3C_CREDENTIALS_URL = "https://www.w3.org/2018/credentials/v1";
-    String VERIFIABLE_PRESENTATION_TYPE = "VerifiablePresentation";
-    String JWS_2020_URL = "https://w3id.org/security/suites/jws-2020/v1";
-    String DID_CONTEXT_URL = "https://www.w3.org/ns/did/v1";
-    String PRESENTATION_SUBMISSION_URL = "https://identity.foundation/presentation-exchange/submission/v1/";
-    String JWS_2020_SIGNATURE_SUITE = "JsonWebSignature2020";
-    String ED25519_SIGNATURE_SUITE = "Ed25519Signature2020"; // not used right now
+import org.eclipse.edc.identitytrust.VcConstants;
+
+public interface IdentityHubConstants extends VcConstants {
 }
