--- spi/identity-hub-spi/src/main/java/org/eclipse/edc/identityhub/spi/model/PresentationQuery.java
@@ -1,78 +1,0 @@
-/*
- *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-package org.eclipse.edc.identityhub.spi.model;
-
-import org.eclipse.edc.identityhub.spi.model.presentationdefinition.PresentationDefinition;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import static org.eclipse.edc.identityhub.spi.model.IdentityHubConstants.IATP_PREFIX;
-
-
-/**
- * Represents a query that is sent to a CredentialService.
- *
- * @see <a href="https://github.com/eclipse-tractusx/identity-trust/blob/main/specifications/M1/verifiable.presentation.protocol.md#411-query-for-presentations">IATP Specification</a>
- */
-public class PresentationQuery {
-    public static final String PRESENTATION_QUERY_SCOPE_PROPERTY = IATP_PREFIX + "scope";
-    public static final String PRESENTATION_QUERY_DEFINITION_PROPERTY = IATP_PREFIX + "presentationDefinition";
-    public static final String PRESENTATION_QUERY_TYPE_PROPERTY = IATP_PREFIX + "Query";
-    private List<String> scopes = new ArrayList<>();
-    private PresentationDefinition presentationDefinition;
-
-    private PresentationQuery() {
-    }
-
-    public List<String> getScopes() {
-        return scopes;
-    }
-
-    public PresentationDefinition getPresentationDefinition() {
-        return presentationDefinition;
-    }
-
-    public static final class Builder {
-        private final PresentationQuery query;
-
-        private Builder() {
-            query = new PresentationQuery();
-        }
-
-        public static Builder newinstance() {
-            return new Builder();
-        }
-
-        public Builder scopes(List<String> scopes) {
-            this.query.scopes = scopes;
-            return this;
-        }
-
-        public Builder scope(String scope) {
-            this.query.scopes.add(scope);
-            return this;
-        }
-
-        public Builder presentationDefinition(PresentationDefinition presentationDefinition) {
-            this.query.presentationDefinition = presentationDefinition;
-            return this;
-        }
-
-        public PresentationQuery build() {
-            return query;
-        }
-    }
-}
