--- spi/identity-hub-spi/src/main/java/org/eclipse/edc/identityhub/spi/model/presentationdefinition/Constraints.java
@@ -1,20 +1,0 @@
-/*
- *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-package org.eclipse.edc.identityhub.spi.model.presentationdefinition;
-
-import java.util.List;
-
-public record Constraints(List<Field> fields) {
-}
